requirements.txt
Django==2.2.10
djangorestframework
psycopg2-binary
redis
PyJWT
django-extensions